+++ test/onyx/triggers/trigger_segment_test.clj
@@ -45,7 +45,7 @@
         state-indices (ws/state-indices event)
         windows-state [(wc/build-window-executor window triggers state-store state-indices task-map)]
         segment1 {:event-time #inst "2016-02-18T12:56:00.910-00:00"}
-        new-segment-event (assoc (t/new-state-event :new-segment event) :segment segment1)
+        new-segment-event (assoc (t/new-state-event :new-segment event 0) :segment segment1)
         ws-1 (ws/fire-state-event windows-state new-segment-event (transient []))
         _ (is (nil? @new-state))
         ws-2 (ws/fire-state-event ws-1 new-segment-event (transient []))